# common/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.0.0-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers (schema and file descriptor)\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `http://MY-REGISTRY-URL/api` by default. Therefore you must prefix all API operation paths with `/api` in this case. For example: `/api/ids/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
        "license": {
...
        }
...
    "paths": {
...
        "/search/artifacts": {
            "summary": "Search for artifacts in the registry.",
            "get": {
                "tags": [
                    "Search",
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "description": "Filter by artifact name.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "offset",
                        "description": "The number of artifacts to skip before starting to collect the result set.  Defaults to 0.",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "limit",
                        "description": "The number of artifacts to return.  Defaults to 20.",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "order",
                        "description": "Sort order, ascending (`asc`) or descending (`desc`).",
                        "schema": {
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "orderby",
                        "description": "The field to sort by.  Can be one of:\n\n* `name`\n* `createdOn`\n",
                        "schema": {
                            "enum": [
                                "name",
                                "createdOn"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "labels",
                        "description": "Filter by label.  Include one or more label to only return artifacts containing all of the\nspecified labels.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "in": "query"
                    },
                    {
                        "name": "properties",
                        "description": "Filter by one or more name/value property.  Separate each name/value pair using a colon.  For\nexample `properties=foo:bar` will return only artifacts with a custom property named `foo`\nand value `bar`.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "in": "query"
                    },
                    {
                        "name": "description",
                        "description": "Filter by description.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "artifactgroup",
                        "description": "Filter by artifact group.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArtifactSearchResults"
                                }
                            }
                        },
                        "description": "On a successful response, returns a result set of artifacts - one for each artifact\nin the registry that matches the criteria."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "searchArtifacts",
                "summary": "Search for artifacts",
                "description": "Returns a paginated list of all artifacts that match the provided filter criteria.\n"
            },
            "post": {
                "tags": [
                    "Search",
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "offset",
                        "description": "The number of artifacts to skip before starting to collect the result set.  Defaults to 0.",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "limit",
                        "description": "The number of artifacts to return.  Defaults to 20.",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "order",
                        "description": "Sort order, ascending (`asc`) or descending (`desc`).",
                        "schema": {
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "orderby",
                        "description": "The field to sort by.  Can be one of:\n\n* `name`\n* `createdOn`\n",
                        "schema": {
                            "enum": [
                                "name",
                                "createdOn"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArtifactSearchResults"
                                }
                            }
                        },
                        "description": "On a successful response, returns a result set of artifacts - one for each artifact\nin the registry that matches the criteria."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "searchArtifactsByContent",
                "summary": "Search for artifacts by content",
                "description": "Returns a paginated list of all artifacts with at least one version that matches the\nposted content.\n"
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Root Type for Error",
                "description": "All error responses, whether `4xx` or `5xx` will include one of these as the response\nbody.",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "The short error message.",
                        "type": "string"
                    },
                    "error_code": {
                        "format": "int32",
                        "description": "The server-side error code.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "Full details about the error.  This might contain a server stack trace, for example.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The error name - typically the classname of the exception thrown by the server.",
                        "type": "string"
                    }
                },
                "example": {
                    "error_code": 500,
                    "message": "An error occurred somewhere."
                }
            },
...
            "ArtifactType": {
                "description": "",
                "enum": [
                    "AVRO",
                    "PROTOBUF",
                    "PROTOBUF_FD",
                    "JSON",
                    "OPENAPI",
                    "ASYNCAPI",
                    "GRAPHQL",
                    "KCONNECT",
                    "WSDL",
                    "XSD",
                    "XML"
                ],
                "type": "string",
                "example": "AVRO",
                "x-codegen-package": "io.apicurio.registry.types"
            },
...
            "ArtifactState": {
                "description": "Describes the state of an artifact or artifact version.  The following states\nare possible:\n\n* ENABLED\n* DISABLED\n* DEPRECATED\n",
                "enum": [
                    "ENABLED",
                    "DISABLED",
                    "DEPRECATED"
                ],
                "type": "string",
                "x-codegen-package": "io.apicurio.registry.types"
            },
            "ArtifactSearchResults": {
                "description": "Describes the response received when searching for artifacts.",
                "required": [
                    "count",
                    "artifacts"
                ],
                "type": "object",
                "properties": {
                    "artifacts": {
                        "description": "The artifacts returned in the result set.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SearchedArtifact"
                        }
                    },
                    "count": {
                        "description": "The total number of artifacts that matched the query that produced the result set (may be \nmore than the number of artifacts in the result set).",
                        "type": "integer"
                    }
                }
            },
...
            "GroupId": {
                "description": "An id of a single Artifact Group.",
                "type": "string",
                "example": "\"my-group\""
            },
            "ArtifactId": {
                "description": "The ID of a single Artifact.",
                "type": "string",
                "example": "\"example-artifact\""
            },
...
            "SearchedArtifact": {
                "description": "Models a single artifact from the result set returned when searching for artifacts.",
                "required": [
                    "createdBy",
                    "createdOn",
                    "id",
                    "type",
                    "state",
                    "group"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/ArtifactId",
                        "description": ""
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "description": {
                        "description": "",
                        "type": "string"
                    },
                    "createdOn": {
                        "format": "date-time",
                        "description": "",
                        "type": "string"
                    },
                    "createdBy": {
                        "description": "",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ArtifactType",
                        "description": ""
                    },
                    "labels": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "state": {
                        "$ref": "#/components/schemas/ArtifactState",
                        "description": ""
                    },
                    "modifiedOn": {
                        "format": "date-time",
                        "description": "",
                        "type": "string"
                    },
                    "modifiedBy": {
                        "description": "",
                        "type": "string"
                    },
                    "groupId": {
                        "$ref": "#/components/schemas/GroupId",
                        "description": ""
                    }
                },
                "example": {
                    "group": "My-Group",
                    "id": "Procurement-Invoice",
                    "name": "Artifact Name",
                    "description": "Description of the artifact",
                    "labels": [
                        "current",
                        "internal"
                    ],
                    "type": "AVRO",
                    "state": "ENABLED",
                    "createdBy": "user1",
                    "createdOn": "2019-03-22T12:51:19Z"
                }
            },
...
        },
        "responses": {
...
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
...
        }
    },
...
}